Changed up how hostname is set, it can be bound to the transport, but the sink will still work
